src/requirements.txt
pydantic
fastapi
pymongo==4.14.0
PyYAML==6.0.2
filetype==1.2.0
python-docx==1.1.2
PyMuPDF==1.24.9
Pillow==10.4.0
pdfplumber==0.11.4
uvicorn
python-multipart